add onSuccess arg for api
